SV22T1020180_Khoi/scripts/chat_gemini_groq/requirements.txt
streamlit>=1.33.0,<2
openai>=1.40.0,<2
google-generativeai>=0.8.0,<1
python-dotenv>=1.0.0,<2
